% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pagespeedonline_objects.R
\name{Result.formattedResults.ruleResults}
\alias{Result.formattedResults.ruleResults}
\title{Result.formattedResults.ruleResults Object}
\usage{
Result.formattedResults.ruleResults()
}
\value{
Result.formattedResults.ruleResults object
}
\description{
Result.formattedResults.ruleResults Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Dictionary of formatted rule results, with one entry for each PageSpeed rule instantiated and run by the server.
}
\seealso{
Other Result functions: \code{\link{Result.formattedResults}},
  \code{\link{Result.pageStats}},
  \code{\link{Result.screenshot}},
  \code{\link{Result.version}}, \code{\link{Result}}
}
