% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pagespeedonline_objects.R
\name{Result.screenshot}
\alias{Result.screenshot}
\title{Result.screenshot Object}
\usage{
Result.screenshot(data = NULL, height = NULL, mime_type = NULL,
  width = NULL)
}
\arguments{
\item{data}{Image data base64 encoded}

\item{height}{Height of screenshot in pixels}

\item{mime_type}{Mime type of image data}

\item{width}{Width of screenshot in pixels}
}
\value{
Result.screenshot object
}
\description{
Result.screenshot Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Base64-encoded screenshot of the page that was analyzed.
}
\seealso{
Other Result functions: \code{\link{Result.formattedResults.ruleResults}},
  \code{\link{Result.formattedResults}},
  \code{\link{Result.pageStats}},
  \code{\link{Result.version}}, \code{\link{Result}}
}
